- Moved xslt and xproc files to subfolders; - Changed the order of the text-to-ssml stages: CSS and normalization come before; - Bug fix in the extract-skippable; - Timeout surrounding the call to the TTS processors; - Renamed initialize() and release() TTSService's method to make things clearer.

+++ ssml-to-audio/src/main/java/org/daisy/pipeline/tts/synthesize/SynthesisWorkerPool.java
@@ -61,9 +61,9 @@ public class SynthesisWorkerPool implements IProgressListener {
 	private Voice mPreviousVoice;
 	private int mNrThreads;
 	private File mAudioDir;
-	private int mTotalTextSize;
-	private int mPrintedProgress;
-	private int mProgress;
+	private long mTotalTextSize;
+	private long mPrintedProgress;
+	private long mProgress;
 
 	public SynthesisWorkerPool(int threadNumber, TTSRegistry registry, AudioEncoder encoder,
 	        IPipelineLogger logger) {
